requirements.txt
pandas
pillow
requests
streamlit
scikit-learn
opencv-python-headless
numpy
streamlit-webrtc
python-dotenv